clevr-bg-api/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
rembg[cpu]>=2.0.50
pillow>=10.0.0
python-multipart>=0.0.6
